[frontend] Sightings header (#11920)
