+++ opencti-platform/opencti-front/src/private/components/events/stix_sighting_relationships/StixSightingRelationship.tsx
@@ -3,12 +3,9 @@ import { useNavigate, useParams } from 'react-router-dom';
 import { graphql } from 'react-relay';
 import Breadcrumbs from 'src/components/Breadcrumbs';
 import { useFormatter } from 'src/components/i18n';
-import { Box, Button, Tooltip } from '@mui/material';
 import Security from 'src/utils/Security';
-import useGranted, { KNOWLEDGE_KNUPDATE, KNOWLEDGE_KNUPDATE_KNDELETE } from 'src/utils/hooks/useGranted';
-import MenuItem from '@mui/material/MenuItem';
-import { useTheme } from '@mui/material/styles';
-import Typography from '@mui/material/Typography';
+import { KNOWLEDGE_KNUPDATE } from 'src/utils/hooks/useGranted';
+import StixSightingRelationshipHeader from '@components/events/stix_sighting_relationships/StixSightingRelationshipHeader';
 import StixSightingRelationshipEdition, { stixSightingRelationshipEditionDeleteMutation } from './StixSightingRelationshipEdition';
 import { commitMutation, defaultCommitMutation, QueryRenderer } from '../../../../relay/environment';
 import { StixSightingRelationshipQuery$data } from './__generated__/StixSightingRelationshipQuery.graphql';
@@ -16,9 +13,6 @@ import Loader from '../../../../components/Loader';
 import StixSightingRelationshipOverview from './StixSightingRelationshipOverview';
 import DeleteDialog from '../../../../components/DeleteDialog';
 import useDeletion from '../../../../utils/hooks/useDeletion';
-import PopoverMenu from '../../../../components/PopoverMenu';
-import type { Theme } from '../../../../components/Theme';
-import { truncate } from '../../../../utils/String';
 
 const stixSightingRelationshipQuery = graphql`
   query StixSightingRelationshipQuery($id: String!) {
@@ -45,8 +39,7 @@ StixSightingRelationshipProps
 > = ({ entityId, paddingRight }) => {
   const { t_i18n } = useFormatter();
   const navigate = useNavigate();
-  const canDelete = useGranted([KNOWLEDGE_KNUPDATE_KNDELETE]);
-  const theme = useTheme<Theme>();
+
   const [editOpen, setEditOpen] = useState<boolean>(false);
   const [openDelete, setOpenDelete] = useState(false);
   const { sightingId } = useParams() as { sightingId: string };
@@ -88,51 +81,11 @@ StixSightingRelationshipProps
                 { label: t_i18n('Sighting'), current: true },
               ]}
               />
-              <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: theme.spacing(3) }}>
-                <div style={{ display: 'flex', alignItems: 'center', gap: theme.spacing(1) }}>
-                  <Tooltip title={headerName}>
-                    <Typography
-                      variant="h1"
-                      sx={{
-                        margin: 0,
-                        lineHeight: 'unset',
-                      }}
-                    >
-                      {truncate(headerName, 80)}
-                    </Typography>
-                  </Tooltip>
-                </div>
-                <div style={{ display: 'flex', alignItems: 'center' }}>
-                  {canDelete && (
-                    <PopoverMenu>
-                      {({ closeMenu }) => (
-                        <Box>
-                          <MenuItem onClick={() => {
-                            handleOpenDelete();
-                            closeMenu();
-                          }}
-                          >
-                            {t_i18n('Delete')}
-                          </MenuItem>
-                        </Box>
-                      )}
-                    </PopoverMenu>
-                  )}
-                  {(
-                    <Security needs={[KNOWLEDGE_KNUPDATE]}>
-                      <Button
-                        variant='contained'
-                        size='medium'
-                        aria-label={t_i18n('Update')}
-                        onClick={handleOpenEdit}
-                        style={{ marginLeft: theme.spacing(0.5) }}
-                      >
-                        {t_i18n('Update')}
-                      </Button>
-                    </Security>
-                    )}
-                </div>
-              </div>
+              <StixSightingRelationshipHeader
+                headerName={headerName}
+                handleOpenEdit={handleOpenEdit}
+                handleOpenDelete={handleOpenDelete}
+              />
               <StixSightingRelationshipOverview
                 entityId={entityId}
                 stixSightingRelationship={result.props.stixSightingRelationship}
